requirements.txt
async-generator==1.10
attrs==22.2.0
black==22.10.0
boto3==1.26.59
botocore==1.29.59
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
debugpy==1.6.4
decorator==4.4.2
demoji==1.1.0
exceptiongroup==1.0.4
flake8==6.0.0
h11==0.14.0
idna==3.4
imageio==2.23.0
imageio-ffmpeg==0.4.7
jmespath==1.0.1
mccabe==0.7.0
mypy-extensions==0.4.3
numpy==1.24.0
outcome==1.2.0
packaging==23.0
pathspec==0.10.2
Pillow==9.3.0
pip-autoremove==0.10.0
pip-review==1.3.0
platformdirs==2.5.4
proglog==0.1.10
pycodestyle==2.10.0
pyflakes==3.0.1
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==0.21.0
pytube==12.1.2
requests==2.28.2
s3transfer==0.6.0
selenium==4.7.2
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
tomli==2.0.1
tqdm==4.64.1
trio==0.22.0
trio-websocket==0.9.2
typing_extensions==4.4.0
urllib3==1.26.13
webdriver-manager==3.8.5
wsproto==1.2.0
